Fix issue: size_t

+++ search_algorithms/0-linear.c
@@ -8,11 +8,11 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	int index = 0;
+	size_t index = 0;
 
 	while (index < size)
 	{
-		printf("Value checked array[%i] = [%i]\n", index, array[index]);
+		printf("Value checked array[%li] = [%i]\n", index, array[index]);
 		if (value == array[index])
 			return (index);
 		index += 1;
